+++ programs/ticket-sale/src/lib.rs
@@ -43,7 +43,7 @@ pub mod ticket_sale {
 	/// * `ticket_type` - The ticket type that will be sold during this sale
 	pub fn create_sale(
 		ctx: Context<CreateSale>,
-		ticket_type_index: u8,
+		ticket_type_index: usize,
 		_cpi_authority_bump: u8,
 		event_id: u64,
 		ticket_type: TicketType,
